[iOS] chore: PincodeManager 파일로 분리
